refactor: hooks in base folder
