+++ src/lib/hooks/index.ts
@@ -1,7 +1,7 @@
-export * from "./common/useCopyToClipboard";
-export * from "./common/useEventListener";
-export * from "./common/useIntersection";
-export * from "./common/useInterval";
-export * from "./common/useLocalStorage";
-export * from "./common/usePrevious";
+export * from "./useCopyToClipboard";
+export * from "./useEventListener";
+export * from "./useIntersection";
+export * from "./useInterval";
+export * from "./useLocalStorage";
+export * from "./usePrevious";
 export * from "./redux";
